# Conflicts:
#	.circleci/config.yml
#	Gemfile.lock
#	fastlane/Fastfile
#	sdk/build.gradle
#	sdk/src/androidTest/java/com/karhoo/sdk/api/AdyenPaymentMethodsIntegrationTest.kt
#	sdk/src/androidTest/java/com/karhoo/sdk/api/AdyenPaymentsIntegrationTest.kt
#	sdk/src/androidTest/java/com/karhoo/sdk/api/QuotesIntegrationTest.kt
#	sdk/src/androidTest/java/com/karhoo/sdk/api/util/ServerRobot.kt
#	sdk/src/main/java/com/karhoo/sdk/api/KarhooError.kt
#	sdk/src/main/java/com/karhoo/sdk/api/model/Availability.kt
#	sdk/src/main/java/com/karhoo/sdk/api/model/AvailabilityVehicle.kt
#	sdk/src/main/java/com/karhoo/sdk/api/model/Constants.kt
#	sdk/src/main/java/com/karhoo/sdk/api/model/PaymentProvider.kt
#	sdk/src/main/java/com/karhoo/sdk/api/model/Provider.kt
#	sdk/src/main/java/com/karhoo/sdk/api/model/QuoteFleet.kt
#	sdk/src/main/java/com/karhoo/sdk/api/model/QuotePrice.kt
#	sdk/src/main/java/com/karhoo/sdk/api/model/QuoteVehicle.kt
#	sdk/src/main/java/com/karhoo/sdk/api/model/adyen/AdyenAmount.kt
#	sdk/src/main/java/com/karhoo/sdk/api/network/annotation/NoAuthorisationHeader.kt
#	sdk/src/main/java/com/karhoo/sdk/api/network/client/APITemplate.kt
#	sdk/src/main/java/com/karhoo/sdk/api/network/request/AdyenPaymentMethodsRequest.kt
#	sdk/src/main/java/com/karhoo/sdk/api/network/request/Passengers.kt
#	sdk/src/main/java/com/karhoo/sdk/api/network/request/QuoteQTA.kt
#	sdk/src/main/java/com/karhoo/sdk/api/network/request/QuoteRequestPoint.kt
#	sdk/src/main/java/com/karhoo/sdk/api/service/availability/AvailabilityInteractor.kt
#	sdk/src/main/java/com/karhoo/sdk/api/service/availability/AvailabilityService.kt
#	sdk/src/main/java/com/karhoo/sdk/api/service/availability/KarhooAvailabilityService.kt
#	sdk/src/main/java/com/karhoo/sdk/api/service/payments/AdyenPaymentMethodsInteractor.kt
#	sdk/src/main/java/com/karhoo/sdk/api/service/payments/AdyenPaymentsInteractor.kt
#	sdk/src/main/java/com/karhoo/sdk/api/service/payments/KarhooPaymentsService.kt
#	sdk/src/main/java/com/karhoo/sdk/api/service/payments/PaymentProviderInteractor.kt
#	sdk/src/main/java/com/karhoo/sdk/api/service/payments/PaymentsService.kt
#	sdk/src/main/java/com/karhoo/sdk/api/service/quotes/KarhooQuotesService.kt
#	sdk/src/main/java/com/karhoo/sdk/api/service/quotes/QuotesService.kt
#	sdk/src/main/java/com/karhoo/sdk/api/service/trips/TripsService.kt
#	sdk/src/main/java/com/karhoo/sdk/api/service/user/UserLoginInteractor.kt
#	sdk/src/test/java/com/karhoo/sdk/api/service/auth/AuthRevokeInteractorTest.kt
#	sdk/src/test/java/com/karhoo/sdk/api/service/availability/AvailabilityInteractorTest.kt
#	sdk/src/test/java/com/karhoo/sdk/api/service/availability/KarhooAvailabilityServiceTest.kt
#	sdk/src/test/java/com/karhoo/sdk/api/service/payments/AdyenPaymentMethodsInteractorTest.kt
#	sdk/src/test/java/com/karhoo/sdk/api/service/payments/AdyenPaymentsInteractorTest.kt
#	sdk/src/test/java/com/karhoo/sdk/api/service/payments/KarhooPaymentsServiceTest.kt
#	sdk/src/test/java/com/karhoo/sdk/api/service/payments/PaymentProviderInteractorTest.kt
#	sdk/src/test/java/com/karhoo/sdk/api/service/quotes/KarhooQuotesServiceTest.kt
#	sdk/src/test/java/com/karhoo/sdk/api/service/trips/BookTripInteractorTest.kt
#	sdk/src/test/java/com/karhoo/sdk/api/service/trips/CancelTripInteractorTest.kt
#	sdk/src/test/java/com/karhoo/sdk/api/service/user/PasswordResetInteractorTest.kt
#	sdk/src/test/java/com/karhoo/sdk/api/service/user/UserLoginInteractorTest.kt

+++ sdk/src/main/java/com/karhoo/sdk/api/model/adyen/AdyenAmount.kt
@@ -7,5 +7,5 @@ import kotlinx.android.parcel.Parcelize
 @Parcelize
 data class AdyenAmount(
         @SerializedName("currency") val currency: String = "",
-        @SerializedName("value") val value: Double = 0.0)
+        @SerializedName("value") val value: Int = 0)
     : Parcelable
